Moved responsibility of tracking current question number to redux store.
The store now trackes questionid, and this is updated through the
changeQuestion action. QuizView has been refactored so that the classes have
been moved to the components folder (and have no knowledge of the redux store)
QuizView is now a container for the Question class.

This marks the first in several coming changes to move functionality to the
redux store.