Further refactor quiz components and added tracking answer in redux store
The bulk of the Question class has been moved to the Quiz class in preparation
for allowing multiple quizzes to choose from. The answers submitted for each
question are now stored in the redux store in pastAnswers, with appropriate
reducer added to append to this.

The selected answer now correctly resets between questions.