Renames action and added QuizListView
Quiz List View is now where quizzes are loaded into the state. It presents a
list of the quizzes in the state and links each to a quiz component when
selected.

Renamed the LOAD_QUIZZES_RECIEVED action to LOAD_QUIZ_RECIEVED as it only
loads one quiz at a time